Refactor loadFile() method in NotesFileParser
